pyproject.toml
[project]
name = "proj-02-django-deploy-test-in-vercel"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.6",
    "djangorestframework>=3.15.2",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "whitenoise>=6.9.0",
]
